Adding database.yml.example, so we have one database file where we can restore from updating invoice generator code which missed a basic check of asserting if invoice exists Updated dashboard layout for first time user Updated acceptance test and added changes to make test pass updated spec_helper and added capybara support files

+++ app/models/invoice_generator.rb
@@ -8,7 +8,7 @@ class InvoiceGenerator
         invoice.order_items << customer.order_items_for_week(weekly_schedule)
       else        
         invoice = weekly_schedule.invoices.find_by_customer_id(customer.id)
-        if invoice.order_items.size != customer.order_items_for_week(weekly_schedule).size
+        if invoice and invoice.order_items.size != customer.order_items_for_week(weekly_schedule).size
           invoice.order_items.delete_all
           invoice.order_items << customer.order_items_for_week(weekly_schedule)
         end
